res templatized csQint() and csQround(). res generasimplifilized SR a wee bit more.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22382 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/software/scan.h
@@ -233,11 +233,11 @@ typedef void (csDrawScanline)
 /// The interface definition for all scan_pi_XXX routines
 typedef void (csDrawPIScanline)
   (void *dest, int len, uint32 *zbuff, int32 u, int32 du, int32 v, int32 dv,
-   uint32 z, int32 dz, unsigned char *bitmap, int bitmap_log2w);
+   float z, float dz, unsigned char *bitmap, int bitmap_log2w);
 /// The interface definition for all scan_pi_XXX_gou_XXX routines
 typedef void (csDrawPIScanlineGouraud)
   (void *dest, int len, uint32 *zbuff, int32 u, int32 du, int32 v, int32 dv,
-   uint32 z, int32 dz, unsigned char *bitmap, int bitmap_log2w,
+   float z, float dz, unsigned char *bitmap, int bitmap_log2w,
    uint32 r, uint32 g, uint32 b, int32 dr, int32 dg, int32 db, bool clamp);
 
 //---//---//---//---//---//---//---//---//---//---//---//---//- Routines //---//
